+++ components/blogs/Blogs.vue
@@ -1,5 +1,5 @@
 <template>
-  <component :is="componentsPrefix + componentName" :content-used-key="contentUsedKey" :blogs-user-role-prop="blogsUserRoleProp" :blogs-uri-prop="blogsUriProp" :create-blog-uri-prop="createBlogUriProp" :authors-uri-prop="authorsUriProp" :categories-uri-prop="categoriesUriProp" :project-id-prop="projectIdProp" :auth-token="authToken" :author-id="authorId" :blogs-user-id-prop="blogsUserIdProp" :blogs-user-name-prop="blogsUserNameProp" :media-translation-prefix="mediaTranslationPrefix" :show-create-blog-button="showCreateBlogButton" :blogs-response-prop="blogsResponseProp" :blog-by-id-uri-prop="blogByIdUriProp" :blog-by-id-response-prop="blogByIdResponseProp" :blog-id-prop="blogId" :create-blog-path="createBlogPath" :edit-blog-path="editBlogPath" :blogs-path="blogsPath" :bo-usage="boUsage" :categories="categories" :server-url="serverUrl" @updateBlogsComponent="onBlogsComponentUpdate" :with-select-blog-button="withSelectBlogButton" :nuxt-sections="nuxtSections" :is-create-blog="isCreateBlog" :blog-id-editing="blogIdEditing" :applied-filters="appliedFilters" @onBlogSelected="(blog) => $emit('getSelectedBlog', blog)"  />
+  <component :is="componentsPrefix + componentName" :content-used-key="contentUsedKey" :blogs-user-role-prop="blogsUserRoleProp" :blogs-uri-prop="blogsUriProp" :create-blog-uri-prop="createBlogUriProp" :authors-uri-prop="authorsUriProp" :categories-uri-prop="categoriesUriProp" :project-id-prop="projectIdProp" :auth-token="authToken" :author-id="authorId" :blogs-user-id-prop="blogsUserIdProp" :blogs-user-name-prop="blogsUserNameProp" :media-translation-prefix="mediaTranslationPrefix" :show-create-blog-button="showCreateBlogButton" :blogs-response-prop="blogsResponseProp" :blog-by-id-uri-prop="blogByIdUriProp" :blog-by-id-response-prop="blogByIdResponseProp" :blog-id-prop="blogId" :default-lang-prop="defaultLang" :create-blog-path="createBlogPath" :edit-blog-path="editBlogPath" :blogs-path="blogsPath" :bo-usage="boUsage" :categories="categories" :server-url="serverUrl" @updateBlogsComponent="onBlogsComponentUpdate" :with-select-blog-button="withSelectBlogButton" :nuxt-sections="nuxtSections" :is-create-blog="isCreateBlog" :blog-id-editing="blogIdEditing" :applied-filters="appliedFilters" @onBlogSelected="(blog) => $emit('getSelectedBlog', blog)"  />
 </template>
 
 <script>
@@ -12,6 +12,7 @@ export default {
     return {
       componentName: 'BlogsListBlogs',
       blogId: '',
+      defaultLang: '',
       authorId: '',
       isCreateBlog: false,
       appliedFilters: '',
@@ -69,6 +70,10 @@ export default {
       default: ""
     },
     blogIdProp: {
+      type: String,
+      default: ""
+    },
+    defaultLangProp: {
       type: String,
       default: ""
     },
@@ -166,6 +171,7 @@ export default {
       this.componentName = 'BlogsEditBlog'
     } else if (this.isEditBlogPath) {
       this.blogId = this.$route.query.id
+      this.defaultLang = this.$route.query.defaultLang
       this.authorId = this.$route.query.userId
       this.componentName = 'BlogsEditBlog'
     }
@@ -181,6 +187,7 @@ export default {
       this.$emit('blogsComponentUpdated', component)
       this.componentName = component.name
       this.blogId = component.blogId
+      this.defaultLang = component.defaultLang
       this.authorId = component.userId
       this.isCreateBlog = component.isCreateBlog
       this.appliedFilters = component.appliedFilters